Refactor.
Remove unnecessary module and codes. Refactor the code.
